refactor: move molecule context
