+++ src/components/StructureExplorer/MoleculeInfo.tsx
@@ -3,7 +3,7 @@ import { AiOutlineLink, AiOutlineSearch } from 'react-icons/ai';
 import { MF } from 'react-mf';
 import { IdcodeSvgRenderer } from 'react-ocl';
 
-import { useMoleculeContext } from '../../hooks/MoleculeContext';
+import { useMoleculeContext } from '../../contexts/molecule_context';
 import { ExtendedWikipediaMolecule } from '../../hooks/fetch_data';
 import Hint from '../Hint';
